Combine leaderboard and entries storage adapters
